requirements.txt
ruff
rich-cli
lmstudio
pydantic
beautifulsoup4
googlesearch-python
